paloalto-autofocus/requirements.txt
pan-python==0.13.0